Map MOT 860 villager trades through UPDATE_TRADE and SAI CraftRecipe.
Java merchants now open from MOT UPDATE_TRADE (window 500), SELECT_TRADE
stores the offer netId, and taking the result encodes CraftRecipe plus
TRADE2 Consume and created-output Take. Bedrock close uses window id -1
so MOT still finds TradeInventory.

+++ src/main/java/net/raphimc/viabedrock/api/model/container/CraftingTableContainer.java
@@ -40,7 +40,7 @@ public class CraftingTableContainer extends Container {
     }
 
     @Override
-    public byte javaContainerId() {
+    public int javaContainerId() {
         return 1;
     }
