+++ src/main/java/net/raphimc/viabedrock/api/util/PacketFactory.java
@@ -189,7 +189,7 @@ public class PacketFactory {
     }
 
     public static void writeJavaContainerSetContent(final PacketWrapper wrapper, final Container container) {
-        wrapper.write(Types.VAR_INT, (int) container.javaContainerId()); // container id
+        wrapper.write(Types.VAR_INT, container.javaContainerId()); // container id
         wrapper.write(Types.VAR_INT, 0); // revision
         wrapper.write(VersionedTypes.V26_1.itemArray, container.getJavaItems()); // items
         wrapper.write(VersionedTypes.V26_1.item, wrapper.user().get(InventoryTracker.class).getHudContainer().getJavaItem(0)); // cursor item
